general cleanup and maintenance (#362)
* general cleanup and maintenance

changed `StatusCode` from `i31` to `u16`.

refactored logic

expanded documentation

added TODOs

made some functions more generic (to improve API w/o causing breaking changes)

* fmt fix

+++ create-rust-app/src/auth/endpoints/service_poem.rs
@@ -16,10 +16,10 @@ use crate::auth::controller::{
 use crate::auth::{controller, Auth, PaginationParams, ID};
 use crate::{Database, Mailer};
 
-fn error_response(status_code: i32, message: &'static str) -> Error {
+fn error_response(status_code: u16, message: &'static str) -> Error {
     Error::from_string(
         json!({ "message": message }).to_string(),
-        StatusCode::from_u16(status_code as u16).unwrap(),
+        StatusCode::from_u16(status_code).unwrap(),
     )
 }
